show posted item(title category created_at)
