+++ resources/js/components/Home.js
@@ -1,5 +1,4 @@
-import React from 'react';
-import ReactDom from 'react-dom';
+import React, { useEffect, useState } from 'react';
 
 
  export class Header extends React.Component {
@@ -37,29 +36,10 @@ import ReactDom from 'react-dom';
   
 
   export class Content extends React.Component {
-    constructor(props){
-      super(props);
-      this.state={"items":["a","b","c","d","e","f","g","h","i"]};
-    }
     render() {
       return (
-        <div className="content">
-                {Object.keys(this.state.items).map(key => (
-            <div className="RecommendedItem">
-              <h1>
-                記事　{this.state.items[key]}
-              </h1>
-              <h5>・おすすめ</h5>
-              <p>おすすめおすすめおすすめおすすめおすすめおすすめおすすめ
-                おすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめ
-                おすすめおすすめおすすめおすすめおすすめおすすめおすすめ
-                おすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめ
-                おすすめおすすめおすすめおすすめおすすめおすすめおすすめ
-                おすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめ
-                おすすめおすすめおすすめおすすめおすすめおすすめおすすめ
-                おすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめおすすめ</p>
-            </div>          
-          ))}
+        <div>
+          <Post />
         </div>
       );  
     }
@@ -78,3 +58,30 @@ import ReactDom from 'react-dom';
     }
   }
   
+  function Post(){
+
+    const [posts, setPosts] = useState([]);
+  
+    useEffect(() => {
+      getPosts()
+    }, [])
+  
+    const getPosts = async () => {
+      const res = await axios.get('/api/');
+      setPosts(res.data.post);
+    }
+
+      return (
+      <div className="content">
+        {posts.map((post) =>
+        <div className="RecommendedItem ">
+          <a href="#" className="content_zone">
+          <h1 key="{post.id}">{post.title}</h1>
+          <p key="{post.id}">category: {post.category}</p>
+          <p key="{post.id}">投稿日: {post.created_at}</p>
+          </a>
+        </div>
+         )}          
+      </div>
+    );
+  }
